Rename package `vectorstores` to `store`
