+++ core/src/commonMain/kotlin/com/xebia/functional/xef/llm/MemoryManagement.kt
@@ -2,7 +2,7 @@ package com.xebia.functional.xef.llm
 
 import com.xebia.functional.xef.conversation.Conversation
 import com.xebia.functional.xef.llm.models.chat.*
-import com.xebia.functional.xef.vectorstores.Memory
+import com.xebia.functional.xef.store.Memory
 import io.ktor.util.date.*
 
 internal object MemoryManagement {
