+++ core/src/commonMain/kotlin/com/xebia/functional/xef/llm/PromptCalculator.kt
@@ -6,7 +6,7 @@ import com.xebia.functional.xef.conversation.Conversation
 import com.xebia.functional.xef.llm.models.chat.Message
 import com.xebia.functional.xef.prompt.Prompt
 import com.xebia.functional.xef.prompt.templates.assistant
-import com.xebia.functional.xef.vectorstores.Memory
+import com.xebia.functional.xef.store.Memory
 
 internal object PromptCalculator {
