+++ core/src/commonMain/kotlin/com/xebia/functional/xef/store/CombinedVectorStore.kt
@@ -1,4 +1,4 @@
-package com.xebia.functional.xef.vectorstores
+package com.xebia.functional.xef.store
 
 import com.xebia.functional.xef.embeddings.Embedding
