+++ core/src/commonMain/kotlin/com/xebia/functional/xef/store/LocalVectorStore.kt
@@ -1,4 +1,4 @@
-package com.xebia.functional.xef.vectorstores
+package com.xebia.functional.xef.store
 
 import arrow.atomic.Atomic
 import arrow.atomic.getAndUpdate
