+++ core/src/commonMain/kotlin/com/xebia/functional/xef/store/Memory.kt
@@ -1,4 +1,4 @@
-package com.xebia.functional.xef.vectorstores
+package com.xebia.functional.xef.store
 
 import com.xebia.functional.xef.llm.models.chat.Message
