+++ core/src/jsMain/kotlin/com/xebia/functional/xef/conversation/JSConversation.kt
@@ -1,7 +1,7 @@
 package com.xebia.functional.xef.conversation
 
-import com.xebia.functional.xef.vectorstores.ConversationId
-import com.xebia.functional.xef.vectorstores.VectorStore
+import com.xebia.functional.xef.store.ConversationId
+import com.xebia.functional.xef.store.VectorStore
 
 class JSConversation(
   override val store: VectorStore,
