+++ core/src/jsMain/kotlin/com/xebia/functional/xef/conversation/PlatformConversation.js.kt
@@ -1,7 +1,7 @@
 package com.xebia.functional.xef.conversation
 
-import com.xebia.functional.xef.vectorstores.ConversationId
-import com.xebia.functional.xef.vectorstores.VectorStore
+import com.xebia.functional.xef.store.ConversationId
+import com.xebia.functional.xef.store.VectorStore
 
 actual abstract class PlatformConversation
 actual constructor(store: VectorStore, conversationId: ConversationId?) : Conversation, AutoClose {
