+++ core/src/jvmMain/kotlin/com/xebia/functional/xef/conversation/JVMConversation.kt
@@ -1,7 +1,7 @@
 package com.xebia.functional.xef.conversation
 
-import com.xebia.functional.xef.vectorstores.ConversationId
-import com.xebia.functional.xef.vectorstores.VectorStore
+import com.xebia.functional.xef.store.ConversationId
+import com.xebia.functional.xef.store.VectorStore
 import java.io.Closeable
 
 open class JVMConversation(
