+++ examples/kotlin/src/main/kotlin/com/xebia/functional/xef/conversation/streaming/OpenAIStreamingExample.kt
@@ -5,7 +5,7 @@ import com.xebia.functional.xef.conversation.llm.openai.OpenAI
 import com.xebia.functional.xef.conversation.llm.openai.OpenAIEmbeddings
 import com.xebia.functional.xef.llm.Chat
 import com.xebia.functional.xef.prompt.Prompt
-import com.xebia.functional.xef.vectorstores.LocalVectorStore
+import com.xebia.functional.xef.store.LocalVectorStore
 
 suspend fun main() {
   val chat: Chat = OpenAI().DEFAULT_CHAT
