+++ gpt4all-kotlin/src/jvmMain/kotlin/com/xebia/functional/gpt4all/Conversation.kt
@@ -1,8 +1,8 @@
 package com.xebia.functional.gpt4all
 
 import com.xebia.functional.xef.conversation.Conversation
-import com.xebia.functional.xef.vectorstores.LocalVectorStore
-import com.xebia.functional.xef.vectorstores.VectorStore
+import com.xebia.functional.xef.store.LocalVectorStore
+import com.xebia.functional.xef.store.VectorStore
 
 suspend inline fun <A> conversation(
   store: VectorStore = LocalVectorStore(HuggingFaceLocalEmbeddings.DEFAULT),
