+++ server/src/main/kotlin/com/xebia/functional/xef/server/services/PersistenceService.kt
@@ -1,7 +1,7 @@
 package com.xebia.functional.xef.server.services
 
 import com.xebia.functional.xef.server.http.routes.Provider
-import com.xebia.functional.xef.vectorstores.VectorStore
+import com.xebia.functional.xef.store.VectorStore
 import io.github.oshai.kotlinlogging.KotlinLogging
 
 abstract class PersistenceService {
